get samples processing out of Nose plugin
